Dependencies enabled for all environments - Phoenix, postgres and other required dependencies allowed in all   environments. - Core components rewritten to use the available dependencies.
