+++ lib/aurora_uix_web/gettext_backend.ex
@@ -1,0 +1,6 @@
+defmodule AuroraUixWeb.GettextBackend do
+  @moduledoc """
+  A fallback module that provides Internationalization with a gettext-based API.
+  """
+  use Gettext.Backend, otp_app: :aurora_uix
+end
